Create a title screen function for mazeMan
Originally there was no title screen and therefore no material to actually initiate the game in the first place. Create a standalone title screen function which consists of the following sequences.

- An intro sequence where the ghost chases the mazeMan across the screen.

- A connecting sequence where the mazeMan moves into the title screen. At this point, display a title printed and a play button.

- An exit sequence triggered by the play button where the ghost chases the mazeMan off screen.

Within the exit sequence, denote the place where the main game loop can be called. Therefore, linking the title screen and the game stage.

As a result, the story of mazeMan begins straight from the the program initiation and transitions smoothly into the actual gameplay.